view all tickets model changes, snake to camel etc

+++ app/Console/Commands/ListenDevEmail.php
@@ -6,8 +6,7 @@ use Illuminate\Console\Command;
 use Webklex\IMAP\Facades\Client;
 use App\Models\Ticket;
 use App\Models\Customer;
-use App\Models\AssignedTicket;
-use App\Models\QueuedTicket;
+use App\Models\Mail;
 use Throwable;
 
 class ListenDevEmail extends Command
@@ -26,13 +25,12 @@ class ListenDevEmail extends Command
         $folder->idle(function ($message) {
             try {
                 $this->info("New message received: " . ($message->subject ?? "No Subject"));
-
+                $this->info("message body" . $message->getTextBody() ?? '');
                 // $message->getFrom()[0]->mail);
 
                 $from = $message->getFrom()[0]->mail;
                 $this->info("From: " . ($from ?? "Unknown"));
                 
-
                 $customerData = [
                     'full_name' => $message->getFrom()[0]->personal ?? "Unknown",
                     'email'     => $message->getFrom()[0]->mail ?? null,
@@ -41,7 +39,6 @@ class ListenDevEmail extends Command
                 if ($customerData['email']) {
                     $customer = Customer::where('email', $customerData['email'])->first();
                     
-
                     if (!$customer) {
                         $customer = Customer::create($customerData);
                     }
@@ -51,25 +48,30 @@ class ListenDevEmail extends Command
                     throw new \Exception("Failed to find or create customer.");
                 }
                 
-                $ticketData = [
-                    'customer_id' => $customer->id,
-                    'subject'       => $message->subject ?? 'No Subject',
+                $mailData = [
+                    'customer_id' => $customer->id,                    
+                    'subject'       => $message->subject,
                     'message'     => $message->getTextBody() ?? '',
                     'message_id'  => $message->message_id ?? ''                    
                 ];
 
-                $ticket = Ticket::create($ticketData);
+                $mail = Mail::create($mailData);
+                if (!$mail || !$mail->id) {
+                    throw new \Exception("Failed to find or create mail.");
+                }
                 $message->delete($expunge = true);
 
-                $queuedTicketData = [
-                    'ticket_id'   => $ticket->id,
-                    'queue_id'    => $ticket->queue_id,                    
-                    'status'      => 0,
-                    'priority'    => 0,
-                    'queue_id'    => 1
+                $ticketData = [
+                    'mail_id' => $mail->id,                                     
+                    'status' => 0,
+                    'priority' => 0,
+                    'queue_id' => 1
                 ];
 
-                QueuedTicket::create($queuedTicketData);
+                $ticket = Ticket::create($ticketData);
+                if (!$ticket || !$ticket->id) {
+                    throw new \Exception("Failed to create ticket.");
+                }
 
             } catch (Throwable $e) {
                 // error_log(json_encode($message));
